+++ src/GeographicPoint.php
@@ -31,6 +31,7 @@ use PHPCoord\CoordinateOperation\ComplexNumber;
 use PHPCoord\CoordinateOperation\ConvertiblePoint;
 use PHPCoord\CoordinateOperation\GeocentricValue;
 use PHPCoord\CoordinateOperation\GeographicValue;
+use PHPCoord\CoordinateOperation\GTXGrid;
 use PHPCoord\CoordinateOperation\IGNFGeocentricTranslationGrid;
 use PHPCoord\CoordinateOperation\NADCON5Grid;
 use PHPCoord\CoordinateOperation\NTv2Grid;
@@ -2170,6 +2171,50 @@ class GeographicPoint extends Point implements ConvertiblePoint
 
         return VerticalPoint::create(
             $this->height->subtract($geoidHeightCorrectionModelFile->getVerticalAdjustment($projected)),
+            $to,
+            $this->getCoordinateEpoch()
+        );
+    }
+
+    /**
+     * Geog3D to Geog2D+GravityRelatedHeight (gtx).
+     */
+    public function geographic3DTo2DPlusGravityHeightGTX(
+        Compound $to,
+        GTXGrid $geoidHeightCorrectionModelFile,
+        string $EPSGCodeForInterpolationCRS
+    ): CompoundPoint {
+        $horizontalPoint = self::create(
+            $this->latitude,
+            $this->longitude,
+            null,
+            $to->getHorizontal(),
+            $this->getCoordinateEpoch()
+        );
+
+        $verticalPoint = VerticalPoint::create(
+            $this->height->subtract($geoidHeightCorrectionModelFile->getAdjustment($this)),
+            $to->getVertical(),
+            $this->getCoordinateEpoch()
+        );
+
+        return CompoundPoint::create(
+            $horizontalPoint,
+            $verticalPoint,
+            $to,
+            $this->getCoordinateEpoch()
+        );
+    }
+
+    /**
+     * Geographic3D to GravityRelatedHeight (gtx).
+     */
+    public function geographic3DToGravityHeightGTX(
+        Vertical $to,
+        GTXGrid $geoidHeightCorrectionModelFile
+    ): VerticalPoint {
+        return VerticalPoint::create(
+            $this->height->subtract($geoidHeightCorrectionModelFile->getAdjustment($this)),
             $to,
             $this->getCoordinateEpoch()
         );